+++ src/processManager/AgentRelaxation.java
@@ -38,8 +38,8 @@ public class AgentRelaxation extends ProcessManager {
 			{
 				//agent.innerSprings();
 				for(Agent neighbour: agents._agentTree.search(
-						(float[]) agent.get("lowerBoundingBox"), /// Add extra margin for pulls!!!
-						(float[]) agent.get("dimensionsBoundingBox"))) 
+						(double[]) agent.get("lowerBoundingBox"), /// Add extra margin for pulls!!!
+						(double[]) agent.get("dimensionsBoundingBox"))) 
 				{
 					if (agent.identity() > neighbour.identity())
 						{
